+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/feature_counter/di/counter_page_with_dependencies.dart
@@ -4,10 +4,12 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_rx_bloc/flutter_rx_bloc.dart';
 import 'package:provider/provider.dart';
 
-import '../../base/data_sources/remote/count_remote_data_source.dart';
-import '../../base/repositories/counter_repository.dart';
 import '../blocs/counter_bloc.dart';
 import '../views/counter_page.dart';
+import '../services/counter_service.dart';
+
+class CounterDependencies {
+  CounterDependencies._(this.context);
 
 class CounterPageWithDependencies extends StatelessWidget {
   const CounterPageWithDependencies({Key? key}) : super(key: key);
@@ -15,26 +17,15 @@ class CounterPageWithDependencies extends StatelessWidget {
   @override
   Widget build(BuildContext context) => MultiProvider(
       providers: [
-        ..._dataSources,
-        ..._repositories,
+        ..._services,
         ..._blocs,
       ],
       child: const CounterPage(),
     );
 
-  /// For your project you should provide a real api in
-  /// lib\base\data_sources\domain\counter\count_remote_data_source.dart
-  /// and data models in lib\base\models and rerun build_runner.
-  List<Provider> get _dataSources => [
-        Provider<CountRemoteDataSource>(
-          create: (context) => CountRemoteDataSource(context.read()),
-        ),
-      ];
-
-  List<Provider> get _repositories => [
-        Provider<CounterRepository>(
-          create: (context) => CounterRepository(
-            context.read(),
+  List<Provider> get _services => [
+        Provider<CounterService>(
+          create: (context) => CounterService(
             context.read(),
           ),
         ),
@@ -42,7 +33,9 @@ class CounterPageWithDependencies extends StatelessWidget {
 
   List<RxBlocProvider> get _blocs => [
         RxBlocProvider<CounterBlocType>(
-          create: (context) => CounterBloc(repository: context.read()),
+          create: (context) => CounterBloc(
+            context.read(),
+          ),
         ),
       ];
 }