+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/feature_counter/views/counter_page.dart
@@ -65,8 +65,8 @@ class CounterPage extends StatelessWidget {
               style: context.designSystem.typography.counterText,
             )
           : Text(
-                snapshot.connectionState.toString(),
-                style: context.designSystem.typography.h2Med16,
+              snapshot.connectionState.toString(),
+              style: context.designSystem.typography.h2Med16,
             );
 
   Widget _buildActionButtons(BuildContext context) =>
@@ -80,6 +80,7 @@ class CounterPage extends StatelessWidget {
               tooltip: context.l10n.featureCounter.increment,
               onPressed: isLoading ? null : bloc.events.increment,
               loading: isLoading && tag == CounterBloc.tagIncrement,
+              heroTag: 'increment',
             ),
             const SizedBox(width: 16),
             ActionButton(
@@ -87,6 +88,7 @@ class CounterPage extends StatelessWidget {
               tooltip: context.l10n.featureCounter.decrement,
               onPressed: isLoading ? null : bloc.events.decrement,
               loading: isLoading && tag == CounterBloc.tagDecrement,
+              heroTag: 'decrement',
             ),
           ],
         ),