+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/feature_login/ui_components/login_form.dart
@@ -89,7 +89,8 @@ class _LoginFormState extends State<LoginForm> {
         ),
       );
 
-  Widget _buildFieldEmail(BuildContext context) => RxTextFormFieldBuilder<LoginBlocType>(
+  Widget _buildFieldEmail(BuildContext context) =>
+      RxTextFormFieldBuilder<LoginBlocType>(
         state: (bloc) => bloc.states.username.translate(context),
         showErrorState: (bloc) => bloc.states.showErrors,
         onChanged: (bloc, value) => bloc.events.setUsername(value),
@@ -103,7 +104,8 @@ class _LoginFormState extends State<LoginForm> {
         ),
       );
 
-  Widget _buildFieldPassword(BuildContext context) => RxTextFormFieldBuilder<LoginBlocType>(
+  Widget _buildFieldPassword(BuildContext context) =>
+      RxTextFormFieldBuilder<LoginBlocType>(
         state: (bloc) => bloc.states.password.translate(context),
         showErrorState: (bloc) => bloc.states.showErrors,
         onChanged: (bloc, value) => bloc.events.setPassword(value),